requirements.txt
openai>=1.40.0
tabulate>=0.9.0
python-dotenv>=1.0.1
pandas 
matplotlib
flask

